initial notes
