+++ homeworks/hw7/index.html
@@ -7,11 +7,7 @@
 </head>
 <body>
     <!-- getPassword() ფუნქციის მიერ დაბრუნებული მნიშვნელობა ჩაწერე 000000-ის მაგივრად -->
-<<<<<<< HEAD
     <div style="display: none;" id="write-password-inside">369576</div>
-=======
-    <div id="write-password-inside">369576</div>
->>>>>>> 637f368 (Add New Post Functionality Tests)
 
     <!-- აქ შეგიძლიათ იმუშავოთ -->
     <div id="app">
