+++ homeworks/hw7/tests.js
@@ -1,34 +1,63 @@
-import { expect } from 'chai';
+import { config, expect } from 'chai';
 import { splitToLines } from "../../src/utils"
 import { specifiers } from './specifiers'
 
 const steps = {
     greetings: 1,
-    console: 2,
-    add_post: 3,
-    add_like: 4,
-    delete_post: 5
+    elements: 2,
+    console: 3,
+    add_post: 4,
+    post_id: 5,
+    add_like: 6,
+    delete_post: 7
 }
 
 export function generateTests(CONFIG){
+    CONFIG.hints = 'on'
 
     CONFIG.isStep(steps.greetings) && describe(`შესავალი`, () => {
-		it(`ამ დავალების მიზანია ჯავასკრიპტის ფუნქციების და DOM-ის ვარჯიში`)
-        
-	})
-
-    CONFIG.isStep(steps.console) && describe(``, () => {
+		it(`ამ დავალებაში რამდენიმე კვირის წინ სემინარზე ნასწავლ მასალაში ვივარჯიშებთ`)
         setPassword(CONFIG)
+        /**
+         * TODO რამის მოფიქრება
+         */
         window.getPassword = () => CONFIG.password
         it(`ამჯერადაც პაროლის ... გახსენი კონსოლი და გამოიძახე getPassword() ფუნქცია`, () => {
 			const n = Number(document.querySelector(`div#write-password-inside`).innerText)
         	expect(n).eql(CONFIG.password)
 		})
+        
 	})
+    
 
+    CONFIG.isStep(steps.elements) && describe('პოსტის დამატება', () => {
+        it(`მოდი დავამატოთ ისეთი სექცია, სადაც იუზერი შეძლებს ტექსტის შეყვანას და შემდეგ დაპოსტვას`)
+        it(`input ელემენტი უკვე ნასწავლი გვაქვს, მაგრამ მისი დამატების შემთხვევაში ვხედავთ 
+        რომ საკმაოდ პატარა ველი არის მრავარპარაგრაფიანი მჭერმეტყველი პოსტებისთვის, რომელიც სოციალურ მედიას ახასიათებს.`)
+        it(`იმისათვის, რომ პატარა ველი არ გვქონდეს, გამოვიყენოთ ერთ-ერთი თაგი textarea, 
+        რომლის id იქნება ${specifiers.textareaId}`, () => {
+            const textareaTag = document.querySelector(`div#app > textarea#${specifiers.textareaId}`)
+            expect(textareaTag).to.not.to.be.null;
+        })
+        it(`<textarea id="new-post-text"></textarea>`)
+
+        it(`textarea-ს ქვემოთ დაამატე ახალი დივ კონტეინერი, სადაც გამოჩნდება იუზერის მიერ დამატებული პოსტები. 
+        კონტეინერის აიდი უნდა იყოს ${specifiers.postsContainerId}`,() => {
+            const postsContainer = document.querySelector(`div#app > div#${specifiers.postsContainerId}`);
+            expect(postsContainer).to.not.to.be.null;
+        })
+
+        /**
+         TODO
+         ღილაკის ელემენტის დამატება (onclick-ის გარეშე)
+         */
+    })
+
+    CONFIG.isStep(steps.console) && describe('DOM ფუნქციონალი', () => {
+        
+    })
+    
     CONFIG.isStep(steps.add_post) && describe('პოსტის დამატება', () => {
-        CONFIG.hints = 'on'
-        it('ის ცვლილებები, რომელიც index.html ფაილში უნდა შეიტანოთ, შეიტანეთ div#app კონტეინერის შიგნით, გარდა სკრიპტის დამატებისა. სკრიპტის დამატებისთვის გადადით index.html ფაილში და იპოვეთ შესაბამისი კომენტარი')
 
         it(`ჯერ სკრიპტ თეგი ჩაამატე. index.js ფაილი უკვე შევქმენი. თქვენს სკრიპტ თეგს id უნდა ჰქონდეს my-script`, () => {
 			let scriptElem = document.querySelector("body > script#my-script")
@@ -38,19 +67,7 @@ export function generateTests(CONFIG){
             expect(srcAttr.includes("index")).to.be.true
 		})
         
-        it(`მოდი დავამატოთ ისეთი სექცია, სადაც იუზერი შეძლებს ტექსტის შეყვანას და შემდეგ დაპოსტვას`)
-        it(`input ელემენტი უკვე ნასწავლი გვაქვს, მაგრამ მისი დამატების შემთხვევაში ვხედავთ რომ საკმაოდ პატარა ველი არის მრავარპარაგრაფიანი მჭერმეტყველი პოსტებისთვის, რომელიც სოციალურ მედიას ახასიათებს.`)
-        it(`იმისათვის, რომ პატარა ველი არ გვქონდეს, გამოვიყენოთ ერთ-ერთი თაგი textarea, რომლის id იქნება ${specifiers.textareaId}`, () => {
-            const textareaTag = document.querySelector(`div#app > textarea#${specifiers.textareaId}`)
-            expect(textareaTag).to.not.to.be.null;
-        })
 
-        it(`ახლა კი დავამატოთ ახალი კონტეინერი, სადაც გამოჩნდება იუზერის მიერ დამატებული პოსტები. კონტეინერის აიდი უნდა იყოს ${specifiers.postsContainerId}`,() => {
-            const postsContainer = document.querySelector(`div#app > div#${specifiers.postsContainerId}`);
-            expect(postsContainer).to.not.to.be.null;
-        })
-
-        it(`დაპოსტვის ფუნქციონალი`)
         it(`გვინდა, რომ ყოველი ახალი პოსტის ელემენტს ქონდეს id, რომელშიც იქნება ნომერი, თუ მერამდენე პოსტია ის. ეს გაგვიმარტივებს პოსტების იდენტიფიცირებას, ვინაიდან შეიძლება არსებობდეს ორი პოსტი, რომელთაც იდენტური ტექსტი აქვთ.`)
         it(`პოსტი ჩვენთვის იქნება div ელემენტი, რომლის class იქნება ${specifiers.postElementClass}. ხოლო მისი id უნდა იწყებოდეს ${specifiers.postElementIdSuffix} და შემდეგ პოსტის იდენტიფიკატორი.`)
         it(`პოსტის ელემენტის შიგნით გვქონდეს კიდევ ერთი div, რომელშიც შევინახავთ პოსტის ტექსტს. ამ div-ის id უნდა იყოს ${specifiers.postElementTextId}`)
@@ -67,6 +84,10 @@ export function generateTests(CONFIG){
             expect(POSTS_ID_COUNTER).eql(1)
         })
 
+        /*
+        TODO: createPostTemplate არის ფუნქცია რომელიც აბრუნებს მნიშვნელობას. ტესტებში ეს ფუნქცია უნდა გამოიძახო
+        და დაბრუნებული მნიშვნელობა შეამოწმო.
+        */
         it(`დავამატოთ ფუნქცია, რომელიც გვიბრუნებს პოსტის template-ს. ეს template უკვე აღვწერეთ წინა ნაბიჯებში როგორი უნდა იყოს. დავარქვათ ამ ფუნცქციას createPostTemplate, გადავცეთ არგუმენტად პოსტის ტექსტი და პოსტის აიდი.`, () => {
             testCreatePostTemplateBasisSection();
             
@@ -93,7 +114,9 @@ export function generateTests(CONFIG){
             
             expect(button.getAttribute("onclick")).eql("newPost()")
         })
-        it("ახლა კი შეგვიძლია newPost() ფუნქცია დავამატოთ. ამ ფუქნციაში, ჯერ წავიკითხოთ რა ჩაწერა იუზერმა textarea-ში. შემდეგ დავამატოთ ახალი პოსტი ამ ტექსტით. შესაბამისი ფუნქცია უკვე დაწერილი გვაქვს, შესაბამისად გამოიძახე createNewPost() ფუნქცია. საბოლოოდ, გავზარდოთ POSTS_ID_COUNTER მნიშვნელობა ერთით", () => {
+        it(`ახლა კი შეგვიძლია newPost() ფუნქცია დავამატოთ. ამ ფუქნციაში, ჯერ წავიკითხოთ რა ჩაწერა იუზერმა textarea-ში. 
+        შემდეგ დავამატოთ ახალი პოსტი ამ ტექსტით. შესაბამისი ფუნქცია უკვე დაწერილი გვაქვს, შესაბამისად გამოიძახე 
+        createNewPost() ფუნქცია. საბოლოოდ, გავზარდოთ POSTS_ID_COUNTER მნიშვნელობა ერთით`, () => {
             // TODO:
             document.getElementById(specifiers.textareaId).value = 'post text 1';
             newPost();
@@ -115,8 +138,17 @@ export function generateTests(CONFIG){
 
 
     })
+    /**
+     * TODO
+     * ამ ნაწილში გადმოვიდეს პოსტის id-ის გაზრდა
+     * არსად არ ამოწმებ რომ addPost ფუნქციის გამოძახების მერე counter გაზრდილია...
+     * 
+     */
+    CONFIG.isStep(steps.post_id) && describe('პოსტებისთვის id-ის მინიჭება', () => {
+        
+    })
 
-    CONFIG.isStep(steps.add_like) && describe('ლაიქის დამატება', () => {
+    CONFIG.isStep(steps.like_button) && describe('ლაიქის დამატება', () => {
         CONFIG.hints = 'on'
         it('ღილაკის დამატება და ფუნქციის მიბმა')
         it(`დაპოსტვის ღილაკისგან განსხვავებით, ლაიქის ღილაკს index.html-ში ვერ დავამატებთ - ყველა პოსტს ცალკე სჭირდება. 
@@ -134,6 +166,13 @@ export function generateTests(CONFIG){
         it('სპანის და ღილაკის ტესტი', () => {
             testCreatePostTemplateLikeSection();
         })
+
+        /**
+         TODO აქ ჯერ შექმნან ცარიელი ფუნქცია და უბრალოდ კონსოლში დაწეროს რომელ პოსტზე დააჭირეს.
+         */
+
+    })
+    CONFIG.isStep(steps.like_function) && describe(`ლაიქის ფუნქცია`, () => {
 
         it(`ახლა დავფიქრდეთ, რა უნდა გააკეთოს likePost ფუნქციამ. 
             ჩვენ გვჭირდება ${specifiers.postLikesNumberClass}-ის innerText-ის შეცვლა, როცა პოსტს დავალაიქებთ. 
@@ -172,42 +211,6 @@ export function generateTests(CONFIG){
         })
     })
 
-    CONFIG.isStep(steps.delete_post) && describe('პოსტის წაშლა', () => {
-        CONFIG.hints = 'on'
-        it('ბოლო ფუნქციონალი რაც დაგვრჩა, არის პოსტის წაშლა')
-        it('ამისთვის, ისევ createPostTemplate ფუნქციის შეცვლა მოგვიწევს, ვინაიდან თითოეულ პოსტს უნდა ჰქონდეს თავისი წაშლის ღილაკი')
-        // TODO: change delete-post hardcoded value
-        it('დაამატე createPostTemplate ფუნქციაში ახალი button, რომლის class იქნება delete-post და onclick ექნება deletePost(${postId}). აქაც, deletePost უნდა გადაეცემოდეს იმ პოსტის აიდი, რომლის წაშლაც გვინდა და უბრალოდ მოხაზულობა აღვწეროთ დაიმპლემენტირებამდე.')
-        it('ღილაკის შემოწმება',() => {
-            testCreatePostTemplateDeletePostSection();
-        })
-        
-        it('საბოლოოდ, დაგვრჩა deletePost ფუნქციის იმპლემენტაცია.')
-        it('ჯერ, იპოვე პოსტის ელემენტი და შეინახე ცვლადში')
-        it('შემდეგ, ცვლადში შეინახე პოსტის მშობელი ელემენტი. მაგალითად, თუ პოსტი შენახულია ცვლადში post, მაშინ მისი მშობელი იქნება post.parentNode')
-        it(`რას ნიშნავს მშობელი ელემენტი? მოდი მაგალითზე დაყრდნობით გავიგოთ.
-            სიტყვაზე, გვაქვს ასეთი html: <div id="parent"> <div id="child"> </div> </div>
-            ამ შემთხვევაში, იმ დივის მშობელი, რომლის id არის child, არის ის div, რომლის id არის parent.
-            ამიტომ, ჩვენ როდესაც მოვძებნით პოსტის ელემენტს, გვეცოდინება, რომ პოსტის მშობელი ელემენტი არის ის div, რომლის id არის ${specifiers.postsContainerId}.
-            ჩვენ კი, ამ კონტეინერიდან გვინდა პოსტის ამოშლა.
-        `)
-        it('ახლა კი, მშობელი ელემენტიდან უნდა წავშალოთ პოსტი. ამისათვის, მშობლის ელემენტით გამოიძახე removeChild ფუნქცია')
-        it('removeChild ფუნქცია უზრუნველყოფს იმას, რომ მშობელი ელემენტიდან ამოიშალოს შვილი ელემენტი. ამ ფუნქციას იმ შვილი ელემენტის DOM-ს მივაწვდით, რომლის წაშლაც გვინდა.')
-        it('parent.removeChild(post)')
-
-        it('deletePost ფუნქციის შემოწმება',() => {
-            setTextareaText('post text 1')
-            newPost()
-
-            let post = document.getElementById('post-1')
-            post.querySelector(`button.${specifiers.postDeleteButtonClass}`).click()
-            
-            post = document.getElementById('post-1')
-            expect(post).to.be.null;
-
-            setTextareaText("")
-        })
-    })
 }
 
 function parseHtmlFromString(htmlString){
@@ -242,16 +245,6 @@ function testCreatePostTemplateLikeSection(){
     expect(postLikeButton.getAttribute('onclick')).eql('likePost(1)')
 }
 
-function testCreatePostTemplateDeletePostSection(){
-    let result = createPostTemplate("post text 1",1);
-    let postElem = parseHtmlFromString(result);
-    testCreatePostTemplateLikeSection();
-
-    let postDeleteButton = postElem.querySelector(`button.${specifiers.postDeleteButtonClass}`);
-    expect(postDeleteButton).to.not.to.be.null;
-    expect(postDeleteButton.getAttribute('class')).eql(specifiers.postDeleteButtonClass);
-    expect(postDeleteButton.getAttribute('onclick')).eql('deletePost(1)')
-}
 
 function setTextareaText(text){
     document.getElementById(specifiers.textareaId).value = text
